Change Navbar Item Creation To Not Use newItem.id
Use the variable "i" to set innerHTML for each newItem that is created in the Navbar.

There may be problems with creating and using IDs for newly created Navbar elements, such as possibly taking precedence over CSS class styling. To prevent this problem, using the upward-counting local variable "i" to reference the nth section HTML (via innerHTML) is intended to do the trick. For now, newItem.id will be commented out, just in case we decide to add it back in later.