Add in console.logs For Debugging smoothScroll()
The smoothScroll() function is intended to scroll the page to the appropriate content section when a Navbar item is clicked.

Currently, there seems to be a problem with or around the "run" y-coordinate being passed into a scrollTo(0, run) function in animation, a function run as part of smoothScroll(). Console.logs are being added in to at least show what code is beginning to run at different points after a user clicks a Navbar item and triggers a "click" eventListener.

A preventDefault method has been applied to prevent the default behavior of jumping straight to the links of content sections when a corresponding Navbar item is clicked. The idea is that the smoothScroll() function will be used instead to allow for a nicer-looking scroll to the relevant section as opposed to an abrupt jump.