Rebuild Anchor Elements as Part of the DOM
Previously, anchor elements as part of the Navbar were created implicitly using innerHTML on their respective parent 'li' elements.

This was problematic as it contributed to DOM-related problems actually relating the anchor elements to the list elements that housed them. As a result of this rebuild of the anchor elements as part of the DOM, the smoothScroll function can now correctly go the relevant section of the page when a Navbar item is clicked. What remains to be done for that function includes building in some easing for scrolling to be less abrupt on-click.