Refactor Scrolling On-Click to Be More Concise
The decision was made to use a simpler scroll function (using scrollIntoView and behavior: smooth).

The desired effect of smooth scrolling when a Navbar item is clicked is still achieved with this simplification, and the code becomes cleaner and more compact. Building out more customization of the scrolling manner in a function may still be desirable in the future, but for now, this far simpler smoothScroll function should work.