Comment Out the Toggling of Active State On Click
Previously, on creation, navbar items would get an event listener added to toggle their active state on click.

When clicked, the navbar item's background would turn light blue. This is being commented out now because the newly added on-scroll listener for the window will toggle the active status of navbar items. Without this change, the two event listeners have a kind of conflict where the togging of the light blue background seems to have a delay, which can also be distracting. Deferring the background-color change to the on-scroll event listener seems to be a good choice for now, especially since clicking a navbar triggers scrolling on the page, and the on-scroll event will be triggered at that point to change the clicked navbar item's background color to light blue.