Add Comments and Toggle Active State of Sections
New comments are added to try to provide more clarity.

Additionally, any page section in a currently active state will now have the "active-section" class. The class is meant to be toggled on or off for page sections whenever scrolling happens in the window, based on their position relative to the viewport.